feat(perception): refactor cone type system - unified yellow vision output with LiDAR size classification
BREAKING CHANGE: Removed ORANGE cone types, replaced with YELLOW_SMALL/YELLOW_BIG

- Vision outputs unified YELLOW (1), LiDAR classifies by size into YELLOW_SMALL/YELLOW_BIG
- Updated all type definitions across vision_core, perception_ros, simulation_core
- Modified fusion logic to classify yellow cones by LiDAR geometry size
- Updated HSV fallback detector with unified yellow detection
- Adjusted confidence threshold from 0.5 to 0.65 to reduce false positives
- Updated track configs: orange/orange_big -> yellow_small/yellow_big
- Fixed planning core color type checks for new yellow types
- Updated visualization config and cone visualizer for new types
- Added backward compatibility for legacy data in cone_visualizer

Refs: vision-lidar-fusion-refactor

+++ src/planning_ros/src/high_speed_tracking/main.cpp
@@ -98,7 +98,7 @@ void doWayMsg(const autodrive_msgs::HUAT_PathLimits &msgs)
     return;
   }
   std::stringstream ss;
-  for (int i = 0; i < msgs.path.size(); i++)
+  for (size_t i = 0; i < msgs.path.size(); i++)
   {
     ss << msgs.path[i].x << "\t" << msgs.path[i].y << "\t" << std::endl;
   }
@@ -123,7 +123,7 @@ void doWayFullMsg(const autodrive_msgs::HUAT_PathLimits &msgs)
     return;
   }
   std::stringstream ss;
-  for (int i = 0; i < msgs.path.size(); i++)
+  for (size_t i = 0; i < msgs.path.size(); i++)
   {
     ss << msgs.path[i].x << "\t" << msgs.path[i].y << "\t" << std::endl;
   }
